# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpaFlow - Spa Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: membershipStatus
          in: query
          schema:
            type: string
            enum: [none, one_time, six_month]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientList"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/memberships:
    post:
      operationId: addMembership
      tags: [clients]
      summary: Add membership to a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipInput"
      responses:
        "201":
          description: Membership added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
  /clients/{id}/rentals:
    get:
      operationId: getClientRentals
      tags: [clients]
      summary: Get a client's rental history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client rental sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalSession"
  /clients/{id}/transactions:
    get:
      operationId: getClientTransactions
      tags: [clients]
      summary: Get a client's transaction history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  # Lockers
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, membershipStatus, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        memberId:
          type: ["string", "null"]
        membershipStatus:
          type: string
          enum: [none, one_time, six_month]
        membershipExpiresAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        # PII fields - only decrypted for MANAGER role
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        documentNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        activeSessions:
          type: array
          items:
            $ref: "#/components/schemas/RentalSession"
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
        address:
          type: string
        documentNumber:
          type: string
        membershipStatus:
          type: string
          enum: [none, one_time, six_month]
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dob:
          type: string
        address:
          type: string
        documentNumber:
          type: string
        membershipStatus:
          type: string
          enum: [none, one_time, six_month]
        notes:
          type: string
    ClientList:
      type: object
      required: [clients, total, page, limit]
      properties:
        clients:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Memberships
    Membership:
      type: object
      required: [id, clientId, type, purchasedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        type:
          type: string
          enum: [one_time, six_month]
        purchasedAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
    MembershipInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [one_time, six_month]
        paymentToken:
          type: string
        idempotencyKey:
          type: string

    # Lockers & Rooms
...
    RentalSession:
      type: object
      required: [id, clientId, resourceType, resourceId, resourceName, status, startTime]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        resourceType:
          type: string
          enum: [locker, room]
        resourceId:
          type: integer
        resourceName:
          type: string
        status:
          type: string
          enum: [active, completed, expired]
        startTime:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        endTime:
          type: ["string", "null"]
          format: date-time
        amountPaid:
          type: ["number", "null"]

    # Pricing
...
    Transaction:
      type: object
      required: [id, clientId, amount, type, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        amount:
          type: number
        tax:
          type: number
        total:
          type: number
        type:
          type: string
          enum: [locker_rental, room_rental, membership, product, renewal, extension]
        squarePaymentId:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...